fix bug in JACK transport relocation
the sync callback did take the first transport position provided by the JACK server when transport is in state JackTransportStarting, requests a relocation internally, and gives its okay to the JACK server.

However, if a client triggers both a relocation and the start of the transport at the same time, the JACK server will provide two cycles in state JackTransportStarting. The first one will still contain the old location and only the second will provide the new one. This makes the `timeline` seeking the first location (wrong) and missing the proper one. At the same the JACK position stored in the client is valid and the BBT data provided as JACK master will be legit (and out of sync with the Non-timeline internal state).

In the provided fix the internal relocation will always and exclusively be triggered if the transport position provided by the JACK server does not match the one stored in the client.